refactor: extract live admission smoke checks
