+++ src/open_range/admit.py
@@ -3,10 +3,8 @@
 from __future__ import annotations
 
 import json
-import shlex
 import shutil
 import subprocess
-import time
 from pathlib import Path
 from typing import Callable, Protocol
 from urllib.parse import urlencode
@@ -27,6 +25,7 @@ from open_range.encryption_enforcement import check_encryption_enforcement
 from open_range.execution import PodActionBackend
 from open_range.identity_enforcement import check_identity_enforcement
 from open_range.k3d_runner import K3dBackend
+from open_range.live_checks import check_live_db_mtls, check_live_service_smoke
 from open_range.mtls_enforcement import check_mtls_enforcement
 from open_range.objectives import evaluate_objective_grader_live
 from open_range.predicates import PredicateEngine
@@ -235,8 +234,8 @@ class LocalAdmissionController:
             snapshot = _ephemeral_snapshot(world, artifacts, reference_bundle)
             backend = PodActionBackend()
             backend.bind(snapshot, release)
-            checks.append(_live_service_smoke_check(world, release))
-            checks.append(_live_db_mtls_check(world, release))
+            checks.append(check_live_service_smoke(world, release))
+            checks.append(check_live_db_mtls(world, release))
             clear_runtime_markers(release, world)
             checks.append(_live_red_reference_check(snapshot, release, backend))
             checks.append(_live_siem_ingest_check(release))
@@ -871,174 +870,6 @@ def _ephemeral_snapshot(
         reference_bundle=reference_bundle,
         world_hash=world_hash(world),
     )
-
-
-def _live_service_smoke_check(world: WorldIR, release) -> ValidatorCheckReport:
-    failures: list[str] = []
-    for service in world.services:
-        runner = _smoke_runner_for_service(world, service.id)
-        cmd = _smoke_probe_command(service)
-        last_error = "smoke failed"
-        ok = False
-        attempts = 10 if service.kind == "db" else 3
-        retry_delay_s = 2.0 if service.kind == "db" else 1.0
-        for attempt in range(attempts):
-            result = run_async(release.pods.exec(runner, cmd, timeout=10.0))
-            if result.ok:
-                ok = True
-                break
-            last_error = result.stderr or result.stdout or "smoke failed"
-            if attempt + 1 < attempts:
-                time.sleep(retry_delay_s)
-        if not ok:
-            failures.append(f"{service.id}:{last_error}")
-    return ValidatorCheckReport(
-        name="live_service_smoke",
-        passed=not failures,
-        details={"failures": failures},
-        error="; ".join(failures),
-    )
-
-
-def _live_db_mtls_check(world: WorldIR, release) -> ValidatorCheckReport:
-    mtls = world.security_runtime.mtls
-    if not mtls or not mtls.get("enabled"):
-        return ValidatorCheckReport(
-            name="live_db_mtls",
-            passed=True,
-            details={"note": "mTLS not enabled"},
-        )
-
-    web_clients = [
-        service
-        for service in world.services
-        if service.kind == "web_app" and "svc-db" in service.dependencies
-    ]
-    if not web_clients:
-        return ValidatorCheckReport(
-            name="live_db_mtls",
-            passed=True,
-            details={"note": "no built-in web client depends on svc-db"},
-        )
-
-    web_client = web_clients[0]
-    positive_cmd = (
-        "mysql --defaults-extra-file=/etc/mysql/conf.d/openrange-client-mtls.cnf "
-        '-Nse "SELECT 1;"'
-    )
-    positive_result = run_async(
-        release.pods.exec(
-            web_client.id,
-            positive_cmd,
-            timeout=15.0,
-            container="db-client-mtls",
-        )
-    )
-
-    negative_cmd = shlex.join(
-        [
-            "mysql",
-            "--protocol=TCP",
-            "--connect-timeout=5",
-            "-h",
-            "svc-db",
-            "-uapp",
-            "-papp-pass",
-            "app",
-            "-Nse",
-            "SELECT 'openrange-db-mtls-ok';",
-        ]
-    )
-    negative_result = run_async(
-        release.pods.exec("sandbox-red", negative_cmd, timeout=15.0)
-    )
-    negative_output = (
-        "\n".join(
-            part for part in (negative_result.stdout, negative_result.stderr) if part
-        )
-        .strip()
-        .lower()
-    )
-
-    failures: list[str] = []
-    if not positive_result.ok:
-        failures.append(
-            f"positive_path:{positive_result.stderr or positive_result.stdout or 'failed'}"
-        )
-    if negative_result.ok:
-        failures.append("no_cert_path:unexpected_success")
-    elif not any(
-        marker in negative_output
-        for marker in (
-            "access denied",
-            "require x509",
-            "x509",
-            "certificate",
-            "ssl",
-            "tls",
-            "1045",
-        )
-    ):
-        failures.append(
-            f"no_cert_path:unexpected_failure_mode:{negative_result.stderr or negative_result.stdout or 'failed'}"
-        )
-
-    return ValidatorCheckReport(
-        name="live_db_mtls",
-        passed=not failures,
-        details={
-            "web_client": web_client.id,
-            "positive_runner": web_client.id,
-            "positive_container": "db-client-mtls",
-            "positive_cmd": positive_cmd,
-            "positive_stdout": positive_result.stdout.strip(),
-            "positive_stderr": positive_result.stderr.strip(),
-            "negative_runner": "sandbox-red",
-            "negative_cmd": negative_cmd,
-            "negative_stdout": negative_result.stdout.strip(),
-            "negative_stderr": negative_result.stderr.strip(),
-        },
-        error="; ".join(failures),
-    )
-
-
-def _smoke_runner_for_service(world: WorldIR, service_id: str) -> str:
-    service_by_id = {service.id: service for service in world.services}
-    host_zone_by_id = {host.id: host.zone for host in world.hosts}
-    service = service_by_id[service_id]
-    zone = host_zone_by_id.get(service.host, "")
-    if zone in {"dmz", "external"}:
-        return "sandbox-red"
-    if zone == "management":
-        return "sandbox-blue"
-    if zone in {"corp", "data"}:
-        runner = _first_green_sandbox_in_zones(world, ("dmz", "corp", zone))
-        if runner:
-            return runner
-    runner = _first_green_sandbox_in_zones(world, (zone,))
-    if runner:
-        return runner
-    return "sandbox-blue" if zone == "management" else "sandbox-red"
-
-
-def _first_green_sandbox_in_zones(world: WorldIR, zones: tuple[str, ...]) -> str:
-    host_zone_by_id = {host.id: host.zone for host in world.hosts}
-    allowed = set(zones)
-    for persona in world.green_personas:
-        zone = host_zone_by_id.get(persona.home_host, "")
-        if zone in allowed:
-            safe = "".join(ch.lower() if ch.isalnum() else "-" for ch in persona.id)
-            return f"sandbox-green-{safe.strip('-')}"
-    return ""
-
-
-def _smoke_probe_command(service: ServiceSpec) -> str:
-    port = service.ports[0] if service.ports else 80
-    if service.id == "svc-web":
-        return f"wget -qO- http://{service.id}:{port}/ | grep -q OpenRange"
-    if service.id == "svc-siem":
-        return "wget -qO- http://svc-siem:9200/all.log >/dev/null"
-    return f"nc -z -w 3 {service.id} {port}"
 
 
 def _live_red_reference_check(
